Add comments and fix failing tests
