+++ cmd/rebuilddb/rebuilddb.go
@@ -79,7 +79,7 @@ func mainCore() int {
 	dbInfo := dcrsqlite.DBInfo{FileName: cfg.DBFileName}
 	//sqliteDB, err := dcrsqlite.InitDB(&dbInfo)
 	sqliteDB, cleanupDB, err := dcrsqlite.InitWiredDB(&dbInfo, nil, client,
-		activeChain, "rebuild_data")
+		activeChain, "rebuild_data", true)
 	defer cleanupDB()
 	if err != nil {
 		log.Errorf("Unable to initialize SQLite database: %v", err)
