+++ db/dbtypes/types.go
@@ -1,15 +1,14 @@
 // Copyright (c) 2018, The Decred developers
 // See LICENSE for details.
+
 package dbtypes
 
 import (
 	"database/sql/driver"
 	"encoding/json"
 	"fmt"
-	"math"
 	"strconv"
 	"strings"
-	"sync"
 
 	"github.com/decred/dcrdata/v3/db/dbtypes/internal"
 )
@@ -94,10 +93,6 @@ const (
 	DayChartGrouping
 	UnknownGrouping
 )
-
-// blockchainSyncStatus defines the status update displayed on the syncing status page
-// when new blocks are being appended into the db.
-var blockchainSyncStatus = new(syncStatus)
 
 const (
 	// InitialDBLoad is a sync where data is first loaded from the chain db into
@@ -566,88 +561,4 @@ type BlockStatus struct {
 	PrevHash    string `json:"previous_hash"`
 	Hash        string `json:"hash"`
 	NextHash    string `json:"next_hash"`
-}
-
-// syncStatus makes it possible to update the user on the progress of the
-// blockchain db syncing that is running in the background after new blocks
-// were detected on system startup.
-type syncStatus struct {
-	sync.RWMutex
-	Update map[string]*SyncStatusInfo
-}
-
-// SyncStatusInfo defines information for a single update type.
-type SyncStatusInfo struct {
-	// percentage change
-	Change float64
-	// used to display the main message about the currect sync.
-	Msg string
-	// displays any other information about the current main sync. This value
-	// may include but not limited to; db indexing, deleting duplicates etc.
-	OtherMsg string
-	// defines estimated time to completing the sync
-	Time       int64
-	UpdateType string
-}
-
-// SyncStatusUpdate defines a thread-safe way to update the sync status updates.
-func SyncStatusUpdate(from, to, timeToComplete int64, updateType, msg string) {
-	blockchainSyncStatus.Lock()
-	defer blockchainSyncStatus.Unlock()
-
-	percentage := 0.0
-	if to > 0 {
-		percentage = math.Floor((float64(from)/float64(to))*10000) / 100
-	}
-
-	val := &SyncStatusInfo{
-		Change:     percentage,
-		Msg:        msg,
-		Time:       timeToComplete,
-		UpdateType: updateType,
-	}
-
-	if len(blockchainSyncStatus.Update) == 0 {
-		blockchainSyncStatus.Update = map[string]*SyncStatusInfo{
-			updateType: val,
-		}
-	} else {
-		blockchainSyncStatus.Update[updateType] = val
-	}
-}
-
-// SyncStatusUpdateOtherMsg is a thread-safe way to update the otherMsg field.
-// This field helps define other tasks that may run in the background and have
-// no feasible way to determine how long they will take to run, e.g its pretty
-// hard to determine how long a given index will take. So it will be indicated
-// just as a task running the background till completion without a progress bar.
-func SyncStatusUpdateOtherMsg(updateType, otherMsg string) {
-	blockchainSyncStatus.Lock()
-	defer blockchainSyncStatus.Unlock()
-
-	val := &SyncStatusInfo{
-		OtherMsg:   otherMsg,
-		UpdateType: updateType,
-	}
-
-	if len(blockchainSyncStatus.Update) == 0 {
-		blockchainSyncStatus.Update = map[string]*SyncStatusInfo{
-			updateType: val,
-		}
-	} else {
-		blockchainSyncStatus.Update[updateType].OtherMsg = otherMsg
-	}
-}
-
-// SyncStatus defines a thread-safe way to read the sync status updates
-func SyncStatus() []*SyncStatusInfo {
-	blockchainSyncStatus.RLock()
-	defer blockchainSyncStatus.RUnlock()
-
-	updates := make([]*SyncStatusInfo, 0)
-	for _, v := range blockchainSyncStatus.Update {
-		updates = append(updates, v)
-	}
-
-	return updates
 }
