+++ db/dcrpg/tables.go
@@ -9,8 +9,8 @@ import (
 	"fmt"
 	"regexp"
 	"strconv"
-	"strings"
 
+	"github.com/decred/dcrdata/v3/explorer"
 	"github.com/decred/dcrdata/v3/db/dbtypes"
 	"github.com/decred/dcrdata/v3/db/dcrpg/internal"
 )
@@ -31,19 +31,6 @@ var createTableStatements = map[string]string{
 var createTypeStatements = map[string]string{
 	"vin_t":  internal.CreateVinType,
 	"vout_t": internal.CreateVoutType,
-}
-
-// indexingInfo defines a minimalistic structure used to append new indexes
-// to be implemented with minimal code duplication.
-type indexingInfo struct {
-	Msg       string
-	IndexFunc func(db *sql.DB) error
-}
-
-// deIndexingInfo defines a minimalistic structure used to append new deindexes
-// to be implemented with minimal code duplication.
-type deIndexingInfo struct {
-	DeIndexFunc func(db *sql.DB) error
 }
 
 // dropDuplicatesInfo defines a minimalistic structure that can be used to
@@ -350,7 +337,7 @@ func (pgb *ChainDB) DeleteDuplicates() error {
 	var err error
 	for _, val := range allDuplicates {
 		msg := fmt.Sprintf("Finding and removing duplicate %s entries...", val.TableName)
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
+		explorer.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
 		log.Info(msg)
 
 		var numRemoved int64
@@ -359,11 +346,11 @@ func (pgb *ChainDB) DeleteDuplicates() error {
 		}
 
 		msg = fmt.Sprintf("Removed %d duplicate %s entries.", numRemoved, val.TableName)
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
+		explorer.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
 		log.Info(msg)
 	}
 	// signal task is done
-	dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, "")
+	explorer.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, "")
 	return nil
 }
 
@@ -394,7 +381,7 @@ func (pgb *ChainDB) DeleteDuplicatesRecovery() error {
 	var err error
 	for _, val := range allDuplicates {
 		msg := fmt.Sprintf("Finding and removing duplicate %s entries...", val.TableName)
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
+		explorer.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
 		log.Info(msg)
 
 		var numRemoved int64
@@ -402,209 +389,11 @@ func (pgb *ChainDB) DeleteDuplicatesRecovery() error {
 			return fmt.Errorf("delete %s duplicate failed: %v", val.TableName, err)
 		}
 
-		msg = fmt.Sprintf("Removed %s duplicate %s entries.", numRemoved, val.TableName)
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
+		msg = fmt.Sprintf("Removed %d duplicate %s entries.", numRemoved, val.TableName)
+		explorer.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, msg)
 		log.Info(msg)
 	}
 	// signal task is done
-	dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, "")
+	explorer.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, "")
 	return nil
-}
-
-// DeindexAll drops all of the indexes in all tables
-func (pgb *ChainDB) DeindexAll() error {
-	allDeIndexes := []deIndexingInfo{
-		// blocks table
-		deIndexingInfo{DeindexBlockTableOnHash},
-		deIndexingInfo{DeindexBlockTableOnHeight},
-
-		// transactions table
-		deIndexingInfo{DeindexTransactionTableOnHashes},
-		deIndexingInfo{DeindexTransactionTableOnBlockIn},
-
-		// vins table
-		deIndexingInfo{DeindexVinTableOnVins},
-		deIndexingInfo{DeindexVinTableOnPrevOuts},
-
-		// vouts table
-		deIndexingInfo{DeindexVoutTableOnTxHashIdx},
-
-		// addresses table
-		deIndexingInfo{DeindexBlockTimeOnTableAddress},
-		deIndexingInfo{DeindexMatchingTxHashOnTableAddress},
-		deIndexingInfo{DeindexAddressTableOnAddress},
-		deIndexingInfo{DeindexAddressTableOnVoutID},
-		deIndexingInfo{DeindexAddressTableOnTxHash},
-
-		// votes table
-		deIndexingInfo{DeindexVotesTableOnCandidate},
-		deIndexingInfo{DeindexVotesTableOnBlockHash},
-		deIndexingInfo{DeindexVotesTableOnHash},
-		deIndexingInfo{DeindexVotesTableOnVoteVersion},
-
-		// misses table
-		deIndexingInfo{DeindexMissesTableOnHash},
-
-		// agendas table
-		deIndexingInfo{DeindexAgendasTableOnBlockTime},
-		deIndexingInfo{DeindexAgendasTableOnAgendaID},
-	}
-
-	var err error
-	for _, val := range allDeIndexes {
-		if err = val.DeIndexFunc(pgb.db); err != nil {
-			warnUnlessNotExists(err)
-			err = nil
-		}
-	}
-
-	if err = pgb.DeindexTicketsTable(); err != nil {
-		warnUnlessNotExists(err)
-		err = nil
-	}
-	return err
-}
-
-// IndexAll creates all of the indexes in all tables
-func (pgb *ChainDB) IndexAll() error {
-	allIndexes := []indexingInfo{
-		// blocks table
-		indexingInfo{Msg: "blocks table on hash", IndexFunc: IndexBlockTableOnHash},
-		indexingInfo{Msg: "blocks table on height", IndexFunc: IndexBlockTableOnHeight},
-
-		// transactions table
-		indexingInfo{Msg: "transactions table on tx/block hashes", IndexFunc: IndexTransactionTableOnHashes},
-		indexingInfo{Msg: "transactions table on block id/indx", IndexFunc: IndexTransactionTableOnBlockIn},
-
-		// vins table
-		indexingInfo{Msg: "vins table on txin", IndexFunc: IndexVinTableOnVins},
-		indexingInfo{Msg: "vins table on prevouts", IndexFunc: IndexVinTableOnPrevOuts},
-
-		// vouts table
-		indexingInfo{Msg: "vouts table on tx hash and index", IndexFunc: IndexVoutTableOnTxHashIdx},
-
-		// votes table
-		indexingInfo{Msg: "votes table on candidate block", IndexFunc: IndexVotesTableOnCandidate},
-		indexingInfo{Msg: "votes table on block hash", IndexFunc: IndexVotesTableOnBlockHash},
-		indexingInfo{Msg: "votes table on block+tx hash", IndexFunc: IndexVotesTableOnHashes},
-		indexingInfo{Msg: "votes table on vote version", IndexFunc: IndexVotesTableOnVoteVersion},
-
-		// misses table
-		indexingInfo{Msg: "misses table", IndexFunc: IndexMissesTableOnHashes},
-
-		// agendas table
-		indexingInfo{Msg: "agendas table on Block Time", IndexFunc: IndexAgendasTableOnBlockTime},
-		indexingInfo{Msg: "agendas table on Agenda ID", IndexFunc: IndexAgendasTableOnAgendaID},
-
-		// Not indexing the address table on vout ID or address here. See
-		// IndexAddressTable to create those indexes.
-		indexingInfo{Msg: "addresses table on tx hash", IndexFunc: IndexAddressTableOnTxHash},
-		indexingInfo{Msg: "addresses table on matching tx hash", IndexFunc: IndexMatchingTxHashOnTableAddress},
-		indexingInfo{Msg: "addresses table on block time", IndexFunc: IndexBlockTimeOnTableAddress},
-	}
-
-	for _, val := range allIndexes {
-		logMsg := "Indexing " + val.Msg + "..."
-		log.Infof(logMsg)
-		if err := val.IndexFunc(pgb.db); err != nil {
-			return err
-		}
-
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, logMsg)
-	}
-	// signal task is done
-	dbtypes.SyncStatusUpdateOtherMsg(dbtypes.InitialDBLoad, "")
-	return nil
-}
-
-// IndexTicketsTable creates the indexes on the tickets table on ticket hash and
-// tx DB ID columns, separately.
-func (pgb *ChainDB) IndexTicketsTable() error {
-	ticketsTableIndexes := []indexingInfo{
-		indexingInfo{Msg: "ticket hash", IndexFunc: IndexTicketsTableOnHashes},
-		indexingInfo{Msg: "ticket pool status", IndexFunc: IndexTicketsTableOnPoolStatus},
-		indexingInfo{Msg: "transaction Db ID", IndexFunc: IndexTicketsTableOnTxDbID},
-	}
-
-	for _, val := range ticketsTableIndexes {
-		logMsg := "Indexing tickets table on " + val.Msg + "..."
-		log.Info(logMsg)
-		if err := val.IndexFunc(pgb.db); err != nil {
-			return err
-		}
-
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.AddressesTableSync, logMsg)
-	}
-	// signal task is done
-	dbtypes.SyncStatusUpdateOtherMsg(dbtypes.AddressesTableSync, "")
-	return nil
-}
-
-// DeindexTicketsTable drops the ticket hash and tx DB ID column indexes for the
-// tickets table.
-func (pgb *ChainDB) DeindexTicketsTable() error {
-	ticketsTablesDeIndexes := []deIndexingInfo{
-		deIndexingInfo{DeindexTicketsTableOnHash},
-		deIndexingInfo{DeindexTicketsTableOnPoolStatus},
-		deIndexingInfo{DeindexTicketsTableOnTxDbID},
-	}
-
-	var err error
-	for _, val := range ticketsTablesDeIndexes {
-		if err = val.DeIndexFunc(pgb.db); err != nil {
-			warnUnlessNotExists(err)
-			err = nil
-		}
-	}
-	return err
-}
-
-func warnUnlessNotExists(err error) {
-	if !strings.Contains(err.Error(), "does not exist") {
-		log.Warn(err)
-	}
-}
-
-// IndexAddressTable creates the indexes on the address table on the vout ID,
-// block_time, matching_tx_hash and address columns, separately.
-func (pgb *ChainDB) IndexAddressTable() error {
-	addressesTableIndexes := []indexingInfo{
-		indexingInfo{Msg: "address", IndexFunc: IndexAddressTableOnAddress},
-		indexingInfo{Msg: "matching tx hash", IndexFunc: IndexMatchingTxHashOnTableAddress},
-		indexingInfo{Msg: "block time", IndexFunc: IndexBlockTimeOnTableAddress},
-		indexingInfo{Msg: "vout Db ID", IndexFunc: IndexAddressTableOnVoutID},
-	}
-
-	for _, val := range addressesTableIndexes {
-		logMsg := "Indexing addresses table on  " + val.Msg + "..."
-		log.Info(logMsg)
-		if err := val.IndexFunc(pgb.db); err != nil {
-			return err
-		}
-
-		dbtypes.SyncStatusUpdateOtherMsg(dbtypes.AddressesTableSync, logMsg)
-	}
-	// signal task is done
-	dbtypes.SyncStatusUpdateOtherMsg(dbtypes.AddressesTableSync, "")
-	return nil
-}
-
-// DeindexAddressTable drops the vin ID, block_time, matching_tx_hash
-// and address column indexes for the address table.
-func (pgb *ChainDB) DeindexAddressTable() error {
-	addressesDeindexes := []deIndexingInfo{
-		deIndexingInfo{DeindexAddressTableOnAddress},
-		deIndexingInfo{DeindexMatchingTxHashOnTableAddress},
-		deIndexingInfo{DeindexBlockTimeOnTableAddress},
-		deIndexingInfo{DeindexAddressTableOnVoutID},
-	}
-
-	var err error
-	for _, val := range addressesDeindexes {
-		if err = val.DeIndexFunc(pgb.db); err != nil {
-			warnUnlessNotExists(err)
-			err = nil
-		}
-	}
-	return err
 }
