+++ views/status.tmpl
@@ -15,7 +15,11 @@
                 <h5>{{.Message}}</h5>
             {{end}}
         </div>
-        <div id="initial-load" class="sync-progress"></div>
+        <div id="initial-load" class="sync-progress">
+        {{if eq .StatusType "Blocks Syncing"}}
+            <img src="/images/loader.gif" style="margin: 0px auto;display: block;"/>
+        {{end}}
+        </div>
         <div id="addresses-sync" class="sync-progress"></div>
     </div>
 {{ template "footer" . }}
